Conflicts:
	integration-tests/webdriver/src/main/java/com/gu/test/pages/Article.java

+++ commercial/app/model/commercial/masterclasses/MasterClassAgent.scala
@@ -17,7 +17,7 @@ object MasterClassAgent extends AdAgent[MasterClass] with ExecutionContexts {
   }
 
   def specificClasses(eventBriteIdStrings: Seq[String]): Seq[MasterClass] = {
-    val eventBriteIds = eventBriteIdStrings map (_.toInt)
+    val eventBriteIds = eventBriteIdStrings map (_.toLong)
     currentAds filter (masterclass => eventBriteIds contains masterclass.eventBriteEvent.id)
   }
